Refactor some module names to align with their content.
