+++ crystal_diffusion/generators/constrained_langevin_generator.py
@@ -4,8 +4,7 @@ import numpy as np
 import torch
 from tqdm import tqdm
 
-from crystal_diffusion.generators.langevin_position_generator import \
-    LangevinGenerator
+from crystal_diffusion.generators.langevin_generator import LangevinGenerator
 from crystal_diffusion.generators.predictor_corrector_position_generator import \
     PredictorCorrectorSamplingParameters
 from crystal_diffusion.models.score_networks.score_network import ScoreNetwork
