+++ tests/generators/test_constrained_langevin_generator.py
@@ -2,9 +2,9 @@ import einops
 import pytest
 import torch
 
-from crystal_diffusion.generators.constrained_langevin_dynamics_generator import (
+from crystal_diffusion.generators.constrained_langevin_generator import (
     ConstrainedLangevinGenerator, ConstrainedLangevinGeneratorParameters)
-from tests.generators.test_langevin_dynamics import \
+from tests.generators.test_langevin_generator import \
     TestAnnealedLangevinDynamics
